pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "WebScraperPortable"
version = "2.0.0"
description = "A self-contained, portable web scraper with session-based output"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0", 
    "pandas>=1.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
full = [
    "ollama>=0.1.0",
    "numpy>=1.21.0", 
    "rich>=13.0.0",
    "python-docx>=0.8.11",
    "PyMuPDF>=1.21.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.10.0",
]

[project.scripts]
webscraper = "WebScraperPortable.__main__:main"

[tool.setuptools.package-dir]
"WebScraperPortable" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = ["-v", "--tb=short"]
filterwarnings = [
    "ignore::DeprecationWarning",
]